requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# F:\Python Projects\Python GUI Projects\PyQt5 Projects\PyQt5 search app\.qt_for_python\uic\Search_ui.py: 11
# F:\Python Projects\Python GUI Projects\PyQt5 Projects\PyQt5 search app\src\main.py: 2,3
PyQt5_sip == 12.8.1

# F:\Python Projects\Python GUI Projects\PyQt5 Projects\PyQt5 search app\src\main.py: 5
wikipedia == 1.4.0
